modified:   kSipP/forms.py 	deleted:    kSipP/scripts/k.py 	new file:   kSipP/scripts/ksipp.py 	modified:   kSipP/static/styles.css 	modified:   kSipP/templates/index.html 	modified:   kSipP/views.py 	modified:   requirements.txt
