+++ kSipP/views.py
@@ -7,58 +7,20 @@ from .forms import configForm, xmlForm, modifyHeaderForm, moreSippOptionsForm, m
 from .forms import xmlUploadForm
 from django.http import HttpResponseBadRequest
 import xml.etree.ElementTree as ET
-import configparser
-import os
-import subprocess
-import psutil
-import signal
-import time
-import re
+import os, time, signal, re
+import subprocess, psutil
+from .scripts.ksipp import get_sipp_processes, fetch_config_data, save_config_data
 from .scripts.kstun import get_ip_info
-
 from .scripts.showXmlFlow import showXmlFlowScript
 from .scripts.modify import modifyHeaderScript, getHeadersFromSipMsgs, tmpXmlBehindNAT, modifynumberxmlpath
-
-# Read initial data from config file
-config_file = os.path.join(str(settings.BASE_DIR), 'config.ini')
-config = configparser.ConfigParser()
-config.read(config_file)
-configd = config['DEFAULT']
-
-#global variables
-def fetch_config_data():
-    # Read initial data from config file
-    config_data = {
-        'uac_remote': configd.get('uac_remote'),
-        'uac_remote_port': configd.get('uac_remote_port'),
-        'uas_remote': configd.get('uas_remote'),
-        'uas_remote_port': configd.get('uas_remote_port'),
-        'local_addr': configd.get('local_addr'),
-        'src_port_uac': configd.get('src_port_uac'),
-        'src_port_uas': configd.get('src_port_uas'),
-        'protocol_uac':configd.get('protocol_uac'),
-        'protocol_uas':configd.get('protocol_uas'),
-        'called_party_number': configd.get('called_party_number'),
-        'calling_party_number': configd.get('calling_party_number'),
-        'total_no_of_calls': configd.get('total_no_of_calls'),
-        'cps': configd.get('cps'),
-        'select_uac':configd.get('select_uac'),
-        'select_uas':configd.get('select_uas'),
-        'stun_server':configd.get('stun_server'),
-
-    }
-    return config_data
-
 
 
 ################### Index Page Functions #############################
 
 @never_cache
 def index(request):
-
     config_data = fetch_config_data()
     
-
     uacXml = f'{config_data["select_uac"]}'
     uasXml = f'{config_data["select_uas"]}'
     uacSrcPort = int(f"{config_data['src_port_uac']}")
@@ -132,16 +94,7 @@ def index(request):
                 config_data['cps'] = moreOptionsForm.cleaned_data['cps']
                 config_data['stun_server'] = moreOptionsForm.cleaned_data['stun_server']
                 
-                # config set for saving in config.ini
-                for configKey, configValue in config_data.items():
-                    config.set('DEFAULT', configKey, str(configValue))
-
-                
-                # Update the config file after config.set
-                ConfigFile = os.path.join(settings.BASE_DIR, 'config.ini')
-                with open(ConfigFile, 'w') as configfile:
-                    config.write(configfile)
-                
+                save_config_data(config_data)
 
 
                 #update script prints on homepage
@@ -180,7 +133,6 @@ def index(request):
 
 
 ######## Index End ############
-
 ######################## Modify XML funtion calls ###############################################
 
 def modifyXml(request):
@@ -300,29 +252,7 @@ def aceXmlEditor(request):
 
 
 
-############# function to get running Sipp Process
 
-def get_sipp_processes():
-    sipp_processes = []
-    sipp_pattern = r"sipp"  # Regular expression to match "sipp" in the command-line
-    shell_name = r"(bash|sh)"  # Regular expression to match shell names
-
-    for process in psutil.process_iter(['pid', 'cmdline']):
-        if process.info['cmdline']:
-            cmdline = ' '.join(process.info['cmdline'])
-            if re.search(sipp_pattern, cmdline) and not re.search(shell_name, cmdline):
-                sipp = os.path.basename(process.info['cmdline'][0])
-                arguments = ' '.join(os.path.basename(arg) if os.path.isabs(arg) else arg for arg in process.info['cmdline'][1:])
-                uac_uas_arg =''.join(os.path.basename(process.info['cmdline'][2]))
-
-                sipp_processes.append({
-                    'pid': process.info['pid'],
-                    'command_line': f"{sipp} {arguments}",
-                    'script_name' : f"{uac_uas_arg}"
-                })
-    sorted_sipp_processes = sorted(sipp_processes, key=lambda x: x['pid'], reverse=True)
-
-    return sorted_sipp_processes
