+++ requirements.txt
@@ -1,4 +1,4 @@
-django==4.2.13
+django==4.2.15
 gunicorn==22.0.0
 lxml==4.9.3
 psutil==5.9.5
